requirements.txt
flask
mysql-connector-python
psycopg2-binary
python-dotenv